Implemented custom exceptions and Repositories classes

+++ HR.LeaveManagement.Application/Features/LeaveAllocations/Handlers/Commands/DeleteLeaveAllocationCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using HR.LeaveManagement.Application.Contracts.Persistence;
+using HR.LeaveManagement.Application.Exceptions;
 using HR.LeaveManagement.Application.Features.LeaveAllocations.Requests.Commands;
 using MediatR;
 
@@ -19,6 +20,8 @@ namespace HR.LeaveManagement.Application.Features.LeaveAllocations.Handlers.Comm
         public async Task<Unit> Handle(DeleteLeaveAllocationCommand request, CancellationToken cancellationToken)
         {
             var leaveAllocation = await leaveAllocationRepository.Get(request.Id);
+            if (leaveAllocation == null)
+                throw new NotFoundException(name: nameof(leaveAllocation), leaveAllocation!.Id);
             await leaveAllocationRepository.Delete(leaveAllocation);
             return Unit.Value;
         }
